[DataRepo] Bound what datareposrc reads by the JSON arrays and the data file
The JSON of a data repository is the element's input, and datareposrc used
its numbers without checking them against each other or against the data
file (items E6 of #4920 and F2 of #4960):

- A tensor_count that decreases wrapped the number of tensors of a sample
  to ~4G, and an index past an array made json-glib return 0 with a
  critical, so the loop allocated 0-byte memories and parsed an 88-byte
  meta header out of each.
- A tensor_size below the 128-byte header was allocated as-is and the
  header was parsed from it, and a huge tensor_size or sample_size went
  straight to the allocator, which aborts the process.
- datareposink parsed the meta header of an upstream memory without
  checking that the memory holds one.

The flexible read path now refuses a sample the JSON does not describe
(its index past sample_offset or tensor_count), a tensor_count that is
past tensor_size (checked before it is narrowed to 32 bits), that
decreases, or that gives more than NNS_TENSOR_SIZE_LIMIT tensors, and a
tensor that is smaller than the meta header or larger than what is left
of the file from the current offset. The last check also covers a sample
offset past the end of the file and bounds what one sample can allocate
by the file size. A negative sample_offset, tensor_count or tensor_size
is refused the same way, so every malformed value reports the sample it
came from instead of a generic stream failure.
A static/media sample larger than the file is refused when the file is
opened. The refusals post STREAM/FORMAT element errors naming the sample.
datareposink uses gst_tensor_meta_info_parse_memory (), which checks the
memory size before parsing.

file_size is widened from gint to guint64: it is compared with 64-bit
offsets now, and a data set larger than 2 GiB would otherwise be refused.

Pipelines that were already broken change behaviour: a JSON whose offset
lies past the file used to end the stream with EOS, and a file shorter
than one sample used to push a partial buffer; both are errors now.

+++ gst/datarepo/gstdatareposrc.h
@@ -45,7 +45,7 @@ struct _GstDataRepoSrc {
   gboolean is_start;            /**< check if datareposrc is started */
   gboolean successful_read;     /**< used for checking EOS when reading more than one images(multi-files) from a path */
   gint fd;                      /**< open file descriptor */
-  gint file_size;               /**< file size, in bytes */
+  guint64 file_size;            /**< file size, in bytes */
   guint64 read_position;        /**< position of fd */
   guint64 fd_offset;            /**< offset of fd */
   guint64 start_offset;         /**< start offset to read */
